% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{addTestDependency}
\alias{addTestDependency}
\title{Allows users to add package dependencies to unit testing, specifically to plot testing}
\usage{
addTestDependency(dep)
}
\arguments{
\item{dep}{A single character value of a package name currently installed on your system}
}
\value{
This function only has a side effect: updating figs/jasp-deps.txt
}
\description{
Testing might fail if dependencies are not the same across platforms (e.g., different versions of jaspGraphs). 
If this is the case then those dependencies should be monitored and errors given when they do not match.
This function allows you to define "unit test breaking" dependencies for plots
}
\examples{

addTestDependency("jaspGraphs")

}
