% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-agent.R
\name{buildAgentTestResult}
\alias{buildAgentTestResult}
\title{Build a rich S3 result object from raw test results.}
\usage{
buildAgentTestResult(allResults, logFile)
}
\description{
Build a rich S3 result object from raw test results.
}
\keyword{internal}
