% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-jasp-snapshots.R
\name{expect_jasp_output}
\alias{expect_jasp_output}
\alias{expect_jasp_table}
\alias{expect_jasp_plot}
\title{JASP Snapshot Testing}
\usage{
expect_jasp_table(x, digits = 10)

expect_jasp_plot(x, name)
}
\arguments{
\item{x}{Code to evaluate.}

\item{digits}{Numerical tolerance; any differences (in the sense of \code{\link[base:all.equal]{base::all.equal()}}) smaller than this value will be ignored.}

\item{name}{The name of the reference plot (a .svg stored in \code{/tests/testthat/_snaps}).}
}
\description{
Wrapper around \code{\link[testthat:expect_snapshot]{testthat::expect_snapshot()}} taking care of testing tables and plots produced by JASP.
}
