% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-agent.R
\name{extractDetailedFailures}
\alias{extractDetailedFailures}
\title{Extract detailed failure messages with backtraces from result list.}
\usage{
extractDetailedFailures(resultList)
}
\description{
Extract detailed failure messages with backtraces from result list.
}
\keyword{internal}
