% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-agent.R
\name{extractDetailedWarnings}
\alias{extractDetailedWarnings}
\title{Extract warning messages from result list.}
\usage{
extractDetailedWarnings(resultList)
}
\description{
Extract warning messages from result list.
}
\keyword{internal}
