% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-agent.R
\name{extractSkipReason}
\alias{extractSkipReason}
\title{Extract skip reason from result list.}
\usage{
extractSkipReason(resultList)
}
\description{
Extract skip reason from result list.
}
\keyword{internal}
