% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-generator.R
\name{generateExampleTestBlock}
\alias{generateExampleTestBlock}
\title{Generate a test block with expectations from analysis results}
\usage{
generateExampleTestBlock(
  analysisName,
  analysisIndex,
  totalAnalyses,
  jaspFileName,
  sourceFolder,
  results
)
}
\arguments{
\item{analysisName}{Name of the analysis function.}

\item{analysisIndex}{Index of this analysis in the JASP file.}

\item{totalAnalyses}{Total number of analyses in the file.}

\item{jaspFileName}{Name of the JASP file.}

\item{sourceFolder}{String indicating the source folder: \code{"library"},
\code{"verified"}, or \code{"other"}.}

\item{results}{The analysis results.}

}
\value{
Character string with the test_that block.
}
\description{
Generate a test block with expectations from analysis results
}
\keyword{internal}
