% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPackage.R
\name{initJaspTools}
\alias{initJaspTools}
\title{Initialize the jaspTools package.}
\usage{
initJaspTools(installJaspPkgs = TRUE, pathJaspRequiredPkgs = NULL,
  pathJaspDesktop = NULL, quiet = FALSE)
}
\arguments{
\item{installJaspPkgs}{Boolean. Should jaspTools install all the JASP modules, jaspBase and jaspGraphs?}

\item{pathJaspRequiredPkgs}{Character path to the root of jasp-required-files if present on the system.}

\item{pathJaspDesktop}{Character path to the root of jasp-desktop if present on the system.}

\item{quiet}{Boolean. Should the installation produce output?}
}
\description{
Ensures that analyses can be run, tested and debugged locally by fetching all of the basic dependencies.
This includes fetching the data library and html files, installing jaspBase and jaspGraphs and adding the required-files.
}
