% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPackage.R
\name{initJaspToolsInteractive}
\alias{initJaspToolsInteractive}
\title{Initialize the jaspTools package in an interactive manner.}
\usage{
initJaspToolsInteractive()
}
\description{
This is a wrapper around \code{initJaspTools} that asks a few questions to determine the desired setup procedure.
Ensures that analyses can be run, tested and debugged locally by fetching all of the basic dependencies.
This includes fetching the data library and html files, installing jaspBase and jaspGraphs and adding the required-files.
}
