% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-setup.R
\name{installJaspPkg}
\alias{installJaspPkg}
\title{Install a JASP R package from jasp-stats}
\usage{
installJaspPkg(pkg, force = FALSE, auth_token = NULL, ...)
}
\arguments{
\item{pkg}{Name of the JASP module (e.g., "jaspBase").}

\item{force}{Boolean. Should the installation overwrite an existing installation if it hasn't changed?}

\item{auth_token}{To install from a private repo, generate a personal access token (PAT) in "https://github.com/settings/tokens" and supply to this argument. This is safer than using a password because you can easily delete a PAT without affecting any others. Defaults to the GITHUB_PAT environment variable.}

\item{...}{Passed on to \code{devtools::install_github}}
}
\description{
Thin wrapper around devtools::install_github.
}
