% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{manageTestPlots}
\alias{manageTestPlots}
\title{Visually inspect failed test plots.}
\usage{
manageTestPlots(name = NULL)
}
\arguments{
\item{name}{Optional string name of the analysis whose plots should be
tested.}
}
\value{
A Shiny app that shows all failed cases. The app allows
test plots to be validated, at which point they are placed in the _snaps
directory and used as a reference for future tests.
}
\description{
This function is a wrapper around \code{testthat::snapshot_review()}. It allows
visual inspection of the plots in the unit tests that produced an error.
If no analysis is specified it will iterate over all test
cases.
}
\examples{

# manageTestPlots("Anova")

}
