% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{monitor}
\alias{monitor}
\title{Tell jaspTools what module you are currently working on.}
\usage{
monitor(modulePaths = ".")
}
\arguments{
\item{modulePaths}{Path to the root of the module (if no path is specified, then the current working directory will be used).}
}
\description{
This information is used to find the correct analysis resources and monitor the module for any changes.
}
\examples{

monitor(c("~/Documents/Github/Regression", "~/Document/Github/Frequencies"))

}
