% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{runTestsTravis}
\alias{runTestsTravis}
\title{Run tests on Travis.}
\usage{
runTestsTravis(modulePath)
}
\arguments{
\item{modulePath}{The path to the module on Travis. Should be obtained from testthat.R.}
}
\description{
Runs all tests that are in the test directory of the module. Should not be used in RStudio.
}
