% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-settings.R
\name{setActiveModule}
\alias{setActiveModule}
\title{Tell jaspTools what module you are currently working on.}
\usage{
setActiveModule(modulePaths = ".", installDeps = TRUE)
}
\arguments{
\item{modulePaths}{Path to the root of the module (if no path is specified, then the current working directory will be used).}

\item{installDeps}{Should jaspTools install missing dependencies of the specified module(s)?}
}
\description{
These module are used to find the R functions etc. in \code{runAnalysis} and the various testing functions.
}
\examples{

setActiveModule(c("~/Documents/Github/Regression", "~/Document/Github/Frequencies"))

}
