% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testModule}
\alias{testModule}
\title{Test all analyses in the currently active module.}
\usage{
testModule(vignettes = TRUE)
}
\arguments{
\item{vignettes}{Would you like to test vignettes as well?}
}
\description{
Tests all R analyses found under modules/tests/testthat. Useful to perform before making
a pull request, to prevent failing builds.
}
