% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{testVignette}
\alias{testVignette}
\alias{testVignettes}
\title{Test JASP Vignettes}
\usage{
testVignette(name)

testVignettes(analysis)
}
\arguments{
\item{name}{Name of the file inside of the \code{vignettes/tests} folder}

\item{analysis}{String name of the analysis to be tested.}
}
\description{
These functions allow you to test vignettes in the \code{vignettes/tests} folder.
\code{testVignette()} tests one specific file, \code{testVignettes()} can test all files in the folder,
or files associated with a specific analysis.
}
\details{
See \link{useTestVignette()} for help on how to create a JASP test vignette.
}
