% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{useTest}
\alias{useTest}
\title{Create or edit test files}
\usage{
useTest(analysis, name = NULL, open = rlang::is_interactive())
}
\arguments{
\item{analysis}{String name of the analysis to create a test file for (case sensitive).}

\item{name}{String name of the test (optional).}

\item{open}{Do you want to open the file automatically in RStudio?}
}
\description{
Both functions mimic \link[usethis]{usethis::use_test()}.
\code{useTest()} creates a "consistency" test file,
\code{useVerifiedTest()} creates a "verified" test file,
meant to be converted into a vignette and a chapter in the jasp-verification-project.
}
