% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{useTestVignette}
\alias{useTestVignette}
\title{Create or edit JASP vignettes}
\usage{
useTestVignette(analysis, name, title = name, open = rlang::is_interactive())
}
\arguments{
\item{analysis}{String name of the analysis that is shown in the vignette (case sensitive).}

\item{name}{String name of the test file.}

\item{title}{String title of the vignette (optional).}

\item{open}{Do you want to open the file automatically in RStudio?}
}
\description{
This function mimics \link[usethis]{usethis::use_article()},
but uses a template for the vignette that is tailored for JASP style vignettes.
This vignette then can be converted into a verified unit test.
}
\details{
Running this function configures the module to have test vignettes.
If necessary, it creates the \code{vignettes/tests} folder, adds \code{knitr} to the \code{Suggests}
field and configures \code{VignetteBuilder} in the \code{DESCRIPTION} file,
adds \code{inst/doc} folder in \code{.gitignore}, and adds \code{vignettes/tests} into \code{.Rbuildignore}.

Then, it creates a file based on the analysis name and name of the test file.

\link{testVignette()} and \link{testVignettes()} can be used for testing the vignette.
}
