% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{view}
\alias{view}
\title{View the tables and plots in a results object.}
\usage{
view(results)
}
\arguments{
\item{results}{A named R list returned from a JASP analysis,
or a json results string copied from the Qt terminal.}
}
\value{
A html page is generated and placed in a temp folder.
}
\description{
\code{view} allows you to view output independently of Qt. It uses the same
javascript/css/html and should generate identical output.
This function may be called directly, but it is more convenient to use \code{JASPTools::run}.
}
