% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-settings.R
\name{viewPkgOptions}
\alias{viewPkgOptions}
\title{View the configuration of jasptools.}
\usage{
viewPkgOptions()
}
\value{
A print of the configurable options.
}
\description{
There are a number of package options you can adjust. These are printed when
you call \code{viewPkgOptions}. Most options are paths to resources, which
should be configured automatically. If the configuration is correct (you
will be notified if it is not) then the only option you should ever change
is .ppi. Setting .ppi to a higher value (through
\code{jasptools::setPkgOption}) results in higher resolution images.
}
