% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-agent.R
\name{writeAgentTestLog}
\alias{writeAgentTestLog}
\title{Write detailed test results to a JSON log file.}
\usage{
writeAgentTestLog(allResults, logFile)
}
\description{
Write detailed test results to a JSON log file.
}
\keyword{internal}
