% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{analysisOptions}
\alias{analysisOptions}
\title{Obtain options to run JASP analyses with.}
\usage{
analysisOptions(source)
}
\arguments{
\item{source}{String containing valid json, or the name of a JASP analysis.
If you provide json, be sure to use single quotes.}
}
\value{
A list containing options you can supply to \code{jasptools::run}.
If \code{source} is an analysis name then all default options have been
filled in and booleans set to FALSE. The options that have no default are
left empty. If \code{hint} is set to TRUE then hints are set for these empty
options; they are placed between \%'s.
}
\description{
\code{analysisOptions} provides an easy way to create analysis options. You
may use the json from the Qt terminal or from the json files found in
resources. The former you have to provide yourself, for the latter you only
have to specify the name of the analysis.
}
\examples{

options <- jasptools::analysisOptions("BinomialTest")
options[["variables"]] <- "contBinom"

# Above and below are identical (below is taken from the Qt terminal)

options <- jasptools::analysisOptions('{
"id" : 0,
"name" : "BinomialTest",
"options" : {
  "VovkSellkeMPR" : false,
  "confidenceInterval" : false,
  "confidenceIntervalInterval" : 0.950,
  "descriptivesPlots" : false,
  "descriptivesPlotsConfidenceInterval" : 0.950,
  "hypothesis" : "notEqualToTestValue",
  "plotHeight" : 300,
  "plotWidth" : 160,
  "testValue" : 0.50,
  "variables" : [ "contBinom" ]
},
"perform" : "run",
"revision" : 0,
"settings" : {
  "ppi" : 192
}
}')

}
