% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-tables.R
\name{expect_equal_tables}
\alias{expect_equal_tables}
\title{Compares JASP tables in unit tests.}
\usage{
expect_equal_tables(test, ref, label = NULL)
}
\arguments{
\item{test}{The table object.}

\item{ref}{The reference table list created previously by applying \code{makeTestTable} to a JASP table returned from \code{runAnalysis}.}

\item{label}{Used to customise failure messages.}
}
\description{
This function compares a stored list of table data, to the table that is created when the tests are run.
}
\examples{

options <- analysisOptions("BinomialTest")
options$variables <- "contBinom"
options$testValue <- 0.6
options$hypothesis <- "greaterThanTestValue"
options$confidenceInterval <- TRUE
options$VovkSellkeMPR <- TRUE
results <- runAnalysis("BinomialTest", "test.csv", options)
table <- results[["results"]][["binomialTable"]][["data"]]
expect_equal_tables(table,
                    list("TRUE", 1, 58, 0, 0.492841460660175, 0.696739870156555, 0.58, 100, 1, "contBinom",
                         "FALSE", 1, 42, 1, 0.336479745077558, 0.999903917924738, 0.42, 100, 1, "contBinom"))

}
