% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-generator.R
\name{generateExampleTestFileContent}
\alias{generateExampleTestFileContent}
\title{Generate test file content for example-based tests}
\usage{
generateExampleTestFileContent(baseName, sanitizedName, testBlocks)
}
\arguments{
\item{baseName}{Original JASP file name without extension.}

\item{sanitizedName}{Sanitized name for use in code.}

\item{testBlocks}{List of test block strings.}
}
\value{
Character string with complete test file content.
}
\description{
Generate test file content for example-based tests
}
\keyword{internal}
