% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-setup.R
\name{installJaspModules}
\alias{installJaspModules}
\title{Install all JASP analysis modules from jasp-stats}
\usage{
installJaspModules(force = FALSE, quiet = FALSE)
}
\arguments{
\item{force}{Boolean. Should JASP reinstall everything or should it only install packages that are not installed on your system?}

\item{quiet}{Boolean. Should the installation procedure produce output?}
}
\description{
This function downloads all JASP modules locally and then installs them, to ensure all dependencies between JASP modules are resolved correctly.
Useful if you're working on modules that have dependencies on other modules.
}
