% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run}
\alias{run}
\title{Run a JASP analysis in R.}
\usage{
run(name, dataset, options, perform = "run", returnResults = FALSE)
}
\arguments{
\item{name}{String indicating the name of the analysis to run.
This name is identical to that of the main function in a JASP analysis.}

\item{dataset}{Data.frame or string; if it's a string it must match one of the JASP datasets (e.g., "debug").}

\item{options}{List of options to supply to the analysis (see also \code{JASPTools::analysisOptions}).}

\item{perform}{String containing either "run" (default) or "init".}

\item{returnResults}{Boolean. Should \code{run} return the analysis results in addition to passing it on to \code{JASPTools::view}?}
}
\description{
\code{run} makes it possible to execute a JASP analysis in R.
Usually this process is a bit cumbersome as there are a number of objects unique to the JASP environment.
Think .ppi, data-reading, etc. These (rcpp) objects are replaced in the JASPTools so you do not have to deal with them.
Note that \code{run} sources JASP analyses every time it runs, so any change in analysis code between calls is incorporated.
The output of the analysis is shown automatically through a call to \code{JASPTools::view}.
}
\examples{
options <- JASPTools::analysisOptions("BinomialTest")
options[["variables"]] <- "contBinom"
JASPTools::run("BinomialTest", "debug", options)
}
