% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-settings.R
\name{setPkgOption}
\alias{setPkgOption}
\title{Change the value of an option in jaspTools.}
\usage{
setPkgOption(name, value)
}
\arguments{
\item{name}{String name of the option.}

\item{value}{Value the option should be set to.}
}
\description{
Sets a package option to a new value (to see what is available run
\code{?viewPkgOptions}). Value changes are automatically
incorporated when any jaspTools function is called.
}
\examples{

setPkgOption("module.dirs", c("~/Documents/Github/Regression", "~/Document/Github/Frequencies"))

}
