% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testAnalysis}
\alias{testAnalysis}
\title{Test a specific JASP analysis.}
\usage{
testAnalysis(name, onlyPlots = FALSE)
}
\arguments{
\item{name}{String name of the analysis to test (case sensitive).}

\item{onlyPlots}{Would you like to only run the tests for plots? This can speed up the generating of reference images in case you are not interested in the other unit tests.}
}
\description{
Tests a specific R analysis found under module/tests/testthat. Useful to perform before
making a pull request, to prevent failing builds.
}
\examples{

testAnalysis("AnovaBayesian")

}
