% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testAnalysis}
\alias{testAnalysis}
\title{Test a specific JASP analysis.}
\usage{
testAnalysis(name)
}
\arguments{
\item{name}{String name of the analysis to test.}
}
\description{
Tests a specific R analysis found under module/tests/testthat. Useful to perform before
making a pull request, to prevent failing builds.
}
\examples{

testAnalysis("AnovaBayesian")

}
