% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{useTest}
\alias{useTest}
\title{Create or edit test files}
\usage{
useTest(analysis, name = NULL, open = rlang::is_interactive())
}
\arguments{
\item{analysis}{String name of the analysis to create a test file for (case sensitive).}

\item{name}{String name of the test (optional).}

\item{open}{Do you want to open the file automatically in RStudio?}
}
\description{
Both functions mimic \code{\link[usethis:use_r]{usethis::use_test()}}.
\code{\link[=useTest]{useTest()}} creates a "consistency" test file,
\code{\link[=useTestVignette]{useTestVignette()}} creates a "verified" test file,
used as a vignette and a verified unit test.
}
