% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-settings.R
\name{viewPkgOptions}
\alias{viewPkgOptions}
\title{View the configuration of jaspTools.}
\usage{
viewPkgOptions()
}
\value{
A print of the configurable options.
}
\description{
There are a number of package options you can adjust. These are printed when
you call \code{viewPkgOptions}. You can adjust these values through
\code{setPkgOption}.
}
\details{
\code{.ppi}:
The pixels per inch in the plots generated by \code{runAnalysis}.

\code{locale}:
The locale that the analysis runs in. Defaults to en_US.UTF-8 to avoid translation of strings.

\code{pkgs.dir}:
The directory where the R packages from jasp-required-files can be found.

\code{data.dir}:
The directory where datafiles can be found so they may be used by \code{runAnalysis}.

\code{html.dir}:
The directory where the javascript, css and .html files can be found to create output identical to JASP. Used by \code{view}.

\code{reinstall.modules}:
When you run an analysis or test it jaspTools calls the *installed* version of the module.
This option specifies if the installed version should be reinstalled automatically when you make any changes to your module.

\code{module.dirs}:
The directories that hold the source for the JASP module(s) you are working on.
These module folders are used to find the R functions etc. in \code{runAnalysis} and the various testing functions.
}
