% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-settings.R
\name{viewPkgOptions}
\alias{viewPkgOptions}
\title{View the configuration of jaspTools.}
\usage{
viewPkgOptions()
}
\value{
A print of the configurable options.
}
\description{
There are a number of package options you can adjust. These are printed when
you call \code{viewPkgOptions}. You can adjust these values through
\code{setPkgOption}.
}
\details{
\code{language}:
The language that the analysis runs in (passed to R environment variable "LANG"). Defaults to "en" to avoid translation of strings.
See https://www.gnu.org/software/gettext/manual/html_node/Usual-Language-Codes.html

\code{data.dirs}:
The directories where datafiles can be found so they may be used by \code{runAnalysis}.

\code{html.dir}:
The directory where the javascript, css and .html files can be found to create output identical to JASP. Used by \code{view}.

\code{view.in.rstudio}:
This option specifies where the html output of \code{runAnalysis} will be shown: inside RStudio or in a separate webbrowser.

\code{reinstall.modules}:
When you run an analysis or test it, jaspTools calls the *installed* version of the module.
This option specifies if the installed version should be reinstalled automatically when you make any changes to your module.

\code{module.dirs}:
The directories that hold the source for the JASP module(s) you are working on.
These module directories are used to find the R functions etc. in \code{runAnalysis} and the various testing functions.
}
